chore(per): nettoyer le potentiel simplifié
