+++ src/features/per/hooks/usePerPotentiel.ts
@@ -6,207 +6,48 @@
  */
 
 import { useState, useCallback, useEffect, useMemo } from 'react';
-import { calculatePerPotentiel } from '../../../engine/per';
+import { calculatePerPotentiel } from '@/engine/per';
 import type {
   PerPotentielInput,
   PerPotentielResult,
   DeclarantRevenus,
   AvisIrPlafonds,
   PerHistoricalBasis,
-} from '../../../engine/per';
-import type { FiscalContext } from '../../../hooks/useFiscalContext';
+} from '@/engine/per';
+import type { FiscalContext } from '@/hooks/useFiscalContext';
 import { getAvisReferenceYears, getPerWorkflowYears } from '../utils/perWorkflowYears';
 import type { PerChildDraft } from '../utils/perParts';
 import { shouldUseProjectionForCalculation } from '../utils/perProjectionScope';
 import { resolvePerCalculationYear } from '../utils/perCalculationYear';
 import { projectionToAvisIrPlafonds } from '../utils/perSyntheticAvis';
-import { getNextPerChildId, normalizePerChildren, normalizePerFoyer } from '../utils/perFoyerState';
 import { buildVisibleSteps } from '../utils/perVisibleSteps';
-import { applyPerSimplifiedPreset, isPerSimplifiedPresetState } from '../utils/perSimplifiedMode';
+import { isPerSimplifiedPresetState } from '../utils/perSimplifiedMode';
+import {
+  buildPerSituationInput,
+  clearPerPotentielSession,
+  loadPerPotentielSession,
+  makeDefaultPerPotentielState,
+  nextPerChildId,
+  normalizePerPotentielState,
+  savePerPotentielSession,
+} from '../utils/perPotentielState';
+import type {
+  PerDeclarantPatch,
+  PerDeclarantScope,
+  PerMode,
+  PerPotentielState,
+  PerSituationPatch,
+  WizardStep,
+} from '../utils/perPotentielState';
 
-const SESSION_KEY = 'ser1:sim:per:potentiel:v4';
-
-export type PerMode = 'versement-n' | 'declaration-n1';
-export type WizardStep = 1 | 2 | 3 | 4 | 5;
-export type PerDeclarantScope = 'revenus-n1' | 'projection-n';
-export type PerDeclarantPatch = { decl: 1 | 2; patch: Partial<DeclarantRevenus> };
-type PerSituationPatch = Partial<{
-  situationFamiliale: 'celibataire' | 'marie';
-  isole: boolean;
-  mutualisationConjoints: boolean;
-}>;
-interface PerPotentielStateOptions {
-  simplifiedMode?: boolean;
-}
-
-export interface PerPotentielState {
-  step: WizardStep;
-  mode: PerMode | null;
-  historicalBasis: PerHistoricalBasis | null;
-  needsCurrentYearEstimate: boolean;
-  avisIr: AvisIrPlafonds | null;
-  avisIr2: AvisIrPlafonds | null;
-  situationFamiliale: 'celibataire' | 'marie';
-  nombreParts: number;
-  isole: boolean;
-  children: PerChildDraft[];
-  projectionSituationFamiliale: 'celibataire' | 'marie';
-  projectionNombreParts: number;
-  projectionIsole: boolean;
-  projectionChildren: PerChildDraft[];
-  projectionMutualisationConjoints: boolean;
-  projectionFoyerEdited: boolean;
-  revenusN1Declarant1: DeclarantRevenus;
-  revenusN1Declarant2: DeclarantRevenus;
-  projectionNDeclarant1: DeclarantRevenus;
-  projectionNDeclarant2: DeclarantRevenus;
-  versementEnvisage: number;
-  mutualisationConjoints: boolean;
-}
-
-const EMPTY_DECLARANT: DeclarantRevenus = {
-  statutTns: false,
-  salaires: 0,
-  fraisReels: false,
-  fraisReelsMontant: 0,
-  art62: 0,
-  bic: 0,
-  retraites: 0,
-  fonciersNets: 0,
-  autresRevenus: 0,
-  cotisationsPer163Q: 0,
-  cotisationsPerp: 0,
-  cotisationsArt83: 0,
-  cotisationsMadelin154bis: 0,
-  cotisationsMadelinRetraite: 0,
-  abondementPerco: 0,
-  cotisationsPrevo: 0,
-};
-
-function normalizeState(state: PerPotentielState, options: PerPotentielStateOptions = {}): PerPotentielState {
-  const baseState = options.simplifiedMode ? applyPerSimplifiedPreset(state) : state;
-  const visibleSteps = buildVisibleSteps(
-    baseState.mode,
-    baseState.historicalBasis,
-    baseState.needsCurrentYearEstimate,
-  );
-  const fallbackStep = visibleSteps[visibleSteps.length - 1] ?? 1;
-  const step = visibleSteps.includes(baseState.step) ? baseState.step : fallbackStep;
-  const foyer = normalizePerFoyer({
-    situationFamiliale: baseState.situationFamiliale,
-    isole: baseState.isole,
-    children: baseState.children,
-    mutualisationConjoints: baseState.mutualisationConjoints,
-  });
-  const projectionFoyer = baseState.projectionFoyerEdited
-    ? normalizePerFoyer({
-      situationFamiliale: baseState.projectionSituationFamiliale,
-      isole: baseState.projectionIsole,
-      children: baseState.projectionChildren,
-      mutualisationConjoints: baseState.projectionMutualisationConjoints,
-    })
-    : foyer;
-
-  return {
-    ...baseState,
-    step,
-    situationFamiliale: foyer.situationFamiliale,
-    isole: foyer.isole,
-    children: foyer.children,
-    mutualisationConjoints: foyer.mutualisationConjoints,
-    nombreParts: foyer.nombreParts,
-    projectionSituationFamiliale: projectionFoyer.situationFamiliale,
-    projectionIsole: projectionFoyer.isole,
-    projectionChildren: projectionFoyer.children,
-    projectionMutualisationConjoints: projectionFoyer.mutualisationConjoints,
-    projectionNombreParts: projectionFoyer.nombreParts,
-  };
-}
-
-function makeDefaultState(options: PerPotentielStateOptions = {}): PerPotentielState {
-  return normalizeState({
-    step: 1,
-    mode: null,
-    historicalBasis: null,
-    needsCurrentYearEstimate: false,
-    avisIr: null,
-    avisIr2: null,
-    situationFamiliale: 'celibataire',
-    nombreParts: 1,
-    isole: false,
-    children: [],
-    projectionSituationFamiliale: 'celibataire',
-    projectionNombreParts: 1,
-    projectionIsole: false,
-    projectionChildren: [],
-    projectionMutualisationConjoints: false,
-    projectionFoyerEdited: false,
-    revenusN1Declarant1: { ...EMPTY_DECLARANT },
-    revenusN1Declarant2: { ...EMPTY_DECLARANT },
-    projectionNDeclarant1: { ...EMPTY_DECLARANT },
-    projectionNDeclarant2: { ...EMPTY_DECLARANT },
-    versementEnvisage: 0,
-    mutualisationConjoints: false,
-  }, options);
-}
-
-function loadSession(options: PerPotentielStateOptions = {}): PerPotentielState | null {
-  try {
-    const raw = sessionStorage.getItem(SESSION_KEY);
-    if (!raw) return null;
-    const parsed = JSON.parse(raw) as Partial<PerPotentielState>;
-    if (!parsed || typeof parsed !== 'object' || !('historicalBasis' in parsed)) {
-      return null;
-    }
-    return normalizeState({
-      ...makeDefaultState(options),
-      ...parsed,
-      revenusN1Declarant1: { ...EMPTY_DECLARANT, ...parsed.revenusN1Declarant1 },
-      revenusN1Declarant2: { ...EMPTY_DECLARANT, ...parsed.revenusN1Declarant2 },
-      projectionNDeclarant1: { ...EMPTY_DECLARANT, ...parsed.projectionNDeclarant1 },
-      projectionNDeclarant2: { ...EMPTY_DECLARANT, ...parsed.projectionNDeclarant2 },
-      children: normalizePerChildren(parsed.children),
-      projectionChildren: normalizePerChildren(parsed.projectionChildren),
-    }, options);
-  } catch {
-    return null;
-  }
-}
-
-function saveSession(state: PerPotentielState): void {
-  try {
-    sessionStorage.setItem(SESSION_KEY, JSON.stringify(state));
-  } catch {
-    /* ignore */
-  }
-}
-
-function buildSituationInput(
-  state: PerPotentielState,
-  scope: PerDeclarantScope,
-  forceCouple = false,
-): PerPotentielInput['situationFiscale'] {
-  const declarant1 = scope === 'revenus-n1' ? state.revenusN1Declarant1 : state.projectionNDeclarant1;
-  const declarant2 = scope === 'revenus-n1' ? state.revenusN1Declarant2 : state.projectionNDeclarant2;
-  const situationFamiliale = scope === 'revenus-n1'
-    ? state.situationFamiliale
-    : state.projectionSituationFamiliale;
-  const nombreParts = scope === 'revenus-n1'
-    ? state.nombreParts
-    : state.projectionNombreParts;
-  const isole = scope === 'revenus-n1'
-    ? state.isole
-    : state.projectionIsole;
-  const isCouple = situationFamiliale === 'marie' || forceCouple;
-
-  return {
-    situationFamiliale,
-    nombreParts,
-    isole,
-    declarant1,
-    declarant2: isCouple ? declarant2 : undefined,
-  };
-}
+export type {
+  PerDeclarantPatch,
+  PerDeclarantScope,
+  PerMode,
+  PerPotentielState,
+  PerSituationPatch,
+  WizardStep,
+} from '../utils/perPotentielState';
 
 export interface UsePerPotentielReturn {
   state: PerPotentielState;
@@ -226,13 +67,8 @@ export interface UsePerPotentielReturn {
   updateProjectionChildMode: (_childId: number, _mode: PerChildDraft['mode']) => void;
   removeChild: (_childId: number) => void;
   removeProjectionChild: (_childId: number) => void;
-  setVersementEnvisage: (_v: number) => void;
-  nextStep: () => void;
-  prevStep: () => void;
   goToStep: (_s: WizardStep) => void;
   reset: () => void;
-  canGoNext: boolean;
-  isCouple: boolean;
 }
 
 export interface UsePerPotentielOptions {
@@ -245,7 +81,7 @@ export function usePerPotentiel(
 ): UsePerPotentielReturn {
   const simplifiedMode = options.simplifiedMode === true;
   const [state, setState] = useState<PerPotentielState>(
-    () => loadSession({ simplifiedMode }) ?? makeDefaultState({ simplifiedMode }),
+    () => loadPerPotentielSession({ simplifiedMode }) ?? makeDefaultPerPotentielState({ simplifiedMode }),
   );
   const years = useMemo(() => getPerWorkflowYears(fiscalContext), [fiscalContext]);
   const visibleSteps = useMemo(
@@ -254,9 +90,9 @@ export function usePerPotentiel(
   );
 
   const persist = useCallback((next: PerPotentielState) => {
-    const normalized = normalizeState(next, { simplifiedMode });
+    const normalized = normalizePerPotentielState(next, { simplifiedMode });
     setState(normalized);
-    saveSession(normalized);
+    savePerPotentielSession(normalized);
   }, [simplifiedMode]);
 
   useEffect(() => {
@@ -265,11 +101,11 @@ export function usePerPotentiel(
     }
 
     setState((previous) => {
-      const normalized = normalizeState(previous, { simplifiedMode: true });
+      const normalized = normalizePerPotentielState(previous, { simplifiedMode: true });
       if (isPerSimplifiedPresetState(previous) && previous.step === normalized.step) {
         return previous;
       }
-      saveSession(normalized);
+      savePerPotentielSession(normalized);
       return normalized;
     });
   }, [simplifiedMode]);
@@ -359,7 +195,7 @@ export function usePerPotentiel(
 
   const addChild = useCallback(() => {
     const nextChild: PerChildDraft = {
-      id: getNextPerChildId(state.children),
+      id: nextPerChildId(state.children),
       mode: 'charge',
     };
     persist({ ...state, children: [...state.children, nextChild] });
@@ -367,7 +203,7 @@ export function usePerPotentiel(
 
   const addProjectionChild = useCallback(() => {
     const nextChild: PerChildDraft = {
-      id: getNextPerChildId(state.projectionChildren),
+      id: nextPerChildId(state.projectionChildren),
       mode: 'charge',
     };
     persist({
@@ -415,23 +251,6 @@ export function usePerPotentiel(
     });
   }, [state, persist]);
 
-  const setVersementEnvisage = useCallback((v: number) => {
-    persist({ ...state, versementEnvisage: v });
-  }, [state, persist]);
-
-  const isCouple = state.situationFamiliale === 'marie';
-  const currentIndex = visibleSteps.indexOf(state.step);
-
-  const nextStep = useCallback(() => {
-    const next = visibleSteps[Math.min(currentIndex + 1, visibleSteps.length - 1)];
-    persist({ ...state, step: next });
-  }, [state, persist, visibleSteps, currentIndex]);
-
-  const prevStep = useCallback(() => {
-    const prev = visibleSteps[Math.max(currentIndex - 1, 0)];
-    persist({ ...state, step: prev });
-  }, [state, persist, visibleSteps, currentIndex]);
-
   const goToStep = useCallback((step: WizardStep) => {
     if (!visibleSteps.includes(step)) {
       return;
@@ -440,30 +259,10 @@ export function usePerPotentiel(
   }, [state, persist, visibleSteps]);
 
   const reset = useCallback(() => {
-    const fresh = makeDefaultState({ simplifiedMode });
+    const fresh = makeDefaultPerPotentielState({ simplifiedMode });
     setState(fresh);
-    try {
-      sessionStorage.removeItem(SESSION_KEY);
-    } catch {
-      /* ignore */
-    }
+    clearPerPotentielSession();
   }, [simplifiedMode]);
-
-  const canGoNext = useMemo(() => {
-    if (state.step !== 1) {
-      return true;
-    }
-
-    if (!state.mode) {
-      return false;
-    }
-
-    if (state.mode === 'versement-n') {
-      return state.historicalBasis !== null;
-    }
-
-    return true;
-  }, [state.step, state.mode, state.historicalBasis]);
 
   const buildInput = useCallback((
     useProjection: boolean,
@@ -495,9 +294,9 @@ export function usePerPotentiel(
         useProjection,
         years,
       }),
-      situationFiscale: buildSituationInput(state, 'revenus-n1'),
+      situationFiscale: buildPerSituationInput(state, 'revenus-n1'),
       projectionFiscale: useProjection
-        ? buildSituationInput(state, 'projection-n', forceProjectionCouple)
+        ? buildPerSituationInput(state, 'projection-n', forceProjectionCouple)
         : undefined,
       avisIr: avisOverride?.avisIr ?? state.avisIr ?? undefined,
       avisIr2: avisOverride?.avisIr2 ?? state.avisIr2 ?? undefined,
@@ -578,12 +377,7 @@ export function usePerPotentiel(
     updateProjectionChildMode,
     removeChild,
     removeProjectionChild,
-    setVersementEnvisage,
-    nextStep,
-    prevStep,
     goToStep,
     reset,
-    canGoNext,
-    isCouple,
   };
 }
